python/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
certifi==2026.2.25
charset-normalizer==3.4.6
click==8.3.1
fastapi==0.135.2
h11==0.16.0
httpx==0.28.1
idna==3.11
iniconfig==2.3.0
numpy==2.4.4
packaging==26.0
pillow==12.1.1
pluggy==1.6.0
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.20.0
pytest==9.0.2
requests==2.33.0
scipy==1.17.1
starlette==1.0.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.42.0
aiohttp==3.11.18